Update encoder modules API
